move: cluster schema to separate file
