+++ redpanda/resources/cluster/resource_cluster.go
@@ -25,18 +25,10 @@ import (
 	controlplanev1beta2 "buf.build/gen/go/redpandadata/cloud/protocolbuffers/go/redpanda/api/controlplane/v1beta2"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/listplanmodifier"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/mapplanmodifier"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
-	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
-	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/redpanda-data/terraform-provider-redpanda/redpanda/cloud"
 	"github.com/redpanda-data/terraform-provider-redpanda/redpanda/config"
 	"github.com/redpanda-data/terraform-provider-redpanda/redpanda/models"
 	"github.com/redpanda-data/terraform-provider-redpanda/redpanda/utils"
-	"github.com/redpanda-data/terraform-provider-redpanda/redpanda/validators"
 )
 
 // Ensure provider defined types fully satisfy framework interfaces.
@@ -81,251 +73,6 @@ func (*Cluster) Schema(_ context.Context, _ resource.SchemaRequest, resp *resour
 	resp.Schema = resourceClusterSchema()
 }
 
-func resourceClusterSchema() schema.Schema {
-	return schema.Schema{
-		Attributes: map[string]schema.Attribute{
-			"name": schema.StringAttribute{
-				Required:    true,
-				Description: "Unique name of the cluster.",
-			},
-			"cluster_type": schema.StringAttribute{
-				Required:      true,
-				Description:   "Cluster type. Type is immutable and can only be set on cluster creation.",
-				Validators:    validators.ClusterTypes(),
-				PlanModifiers: []planmodifier.String{stringplanmodifier.RequiresReplace()},
-			},
-			"connection_type": schema.StringAttribute{
-				Required:      true,
-				Description:   "Cluster connection type. Private clusters are not exposed to the internet. For BYOC clusters, Private is best-practice.",
-				Validators:    validators.ConnectionTypes(),
-				PlanModifiers: []planmodifier.String{stringplanmodifier.RequiresReplace()},
-			},
-			"cloud_provider": schema.StringAttribute{
-				Optional:      true,
-				Description:   "Cloud provider where resources are created.",
-				PlanModifiers: []planmodifier.String{stringplanmodifier.RequiresReplace()},
-				Validators:    validators.CloudProviders(),
-			},
-			"redpanda_version": schema.StringAttribute{
-				Optional:      true,
-				Description:   "Current Redpanda version of the cluster.",
-				PlanModifiers: []planmodifier.String{stringplanmodifier.RequiresReplace()},
-			},
-			"throughput_tier": schema.StringAttribute{
-				Required:      true,
-				Description:   "Throughput tier of the cluster.",
-				PlanModifiers: []planmodifier.String{stringplanmodifier.RequiresReplace()},
-			},
-			"region": schema.StringAttribute{
-				Optional:      true,
-				Description:   "Cloud provider region. Region represents the name of the region where the cluster will be provisioned.",
-				PlanModifiers: []planmodifier.String{stringplanmodifier.RequiresReplace()},
-			},
-			"zones": schema.ListAttribute{
-				Optional:      true,
-				Description:   "Zones of the cluster. Must be valid zones within the selected region. If multiple zones are used, the cluster is a multi-AZ cluster.",
-				ElementType:   types.StringType,
-				PlanModifiers: []planmodifier.List{listplanmodifier.RequiresReplace()},
-			},
-			"allow_deletion": schema.BoolAttribute{
-				Optional:    true,
-				Description: "Allows deletion of the cluster. Defaults to true. Should probably be set to false for production use.",
-			},
-			"tags": schema.MapAttribute{
-				Optional:      true,
-				Description:   "Tags placed on cloud resources. If the cloud provider is GCP and the name of a tag has the prefix \"gcp.network-tag.\", the tag is a network tag that will be added to the Redpanda cluster GKE nodes. Otherwise, the tag is a normal tag. For example, if the name of a tag is \"gcp.network-tag.network-tag-foo\", the network tag named \"network-tag-foo\" will be added to the Redpanda cluster GKE nodes. Note: The value of a network tag will be ignored. See the details on network tags at https://cloud.google.com/vpc/docs/add-remove-network-tags.",
-				ElementType:   types.StringType,
-				PlanModifiers: []planmodifier.Map{mapplanmodifier.RequiresReplace()},
-			},
-			"resource_group_id": schema.StringAttribute{
-				Required:      true,
-				Description:   "Resource group ID of the cluster.",
-				PlanModifiers: []planmodifier.String{stringplanmodifier.RequiresReplace()},
-			},
-			"network_id": schema.StringAttribute{
-				Required:      true,
-				Description:   "Network ID where cluster is placed.",
-				PlanModifiers: []planmodifier.String{stringplanmodifier.RequiresReplace()},
-			},
-			"id": schema.StringAttribute{
-				Computed:      true,
-				Description:   "ID of the cluster. ID is an output from the Create Cluster endpoint and cannot be set by the caller.",
-				PlanModifiers: []planmodifier.String{stringplanmodifier.UseStateForUnknown()},
-			},
-			"cluster_api_url": schema.StringAttribute{
-				Computed:      true,
-				Description:   "The URL of the cluster API.",
-				PlanModifiers: []planmodifier.String{stringplanmodifier.UseStateForUnknown()},
-			},
-			"aws_private_link": schema.SingleNestedAttribute{
-				Optional:    true,
-				Description: "The AWS Private Link configuration.",
-				Attributes: map[string]schema.Attribute{
-					"enabled": schema.BoolAttribute{
-						Required:    true,
-						Description: "Whether Redpanda AWS Private Link Endpoint Service is enabled.",
-					},
-					"connect_console": schema.BoolAttribute{
-						Required:    true,
-						Description: "Whether Console is connected in Redpanda AWS Private Link Service.",
-					},
-					"allowed_principals": schema.ListAttribute{
-						ElementType: types.StringType,
-						Required:    true,
-						Description: "The ARN of the principals that can access the Redpanda AWS PrivateLink Endpoint Service. To grant permissions to all principals, use an asterisk (*).",
-					},
-				},
-				Validators: []validator.Object{
-					validators.CloudProviderDependentValidator{
-						AttributeName: "aws_private_link",
-						CloudProvider: "aws",
-					},
-				},
-			},
-			"azure_private_link": schema.SingleNestedAttribute{
-				Optional:    true,
-				Description: "The Azure Private Link configuration.",
-				Attributes: map[string]schema.Attribute{
-					"allowed_subscriptions": schema.ListAttribute{
-						ElementType: types.StringType,
-						Required:    true,
-						Description: "The subscriptions that can access the Redpanda Azure PrivateLink Endpoint Service. To grant permissions to all principals, use an asterisk (*).",
-					},
-					"connect_console": schema.BoolAttribute{
-						Required:    true,
-						Description: "Whether Console is connected in Redpanda Azure Private Link Service.",
-					},
-					"enabled": schema.BoolAttribute{
-						Required:    true,
-						Description: "Whether Redpanda Azure Private Link Endpoint Service is enabled.",
-					},
-				},
-				Validators: []validator.Object{
-					validators.CloudProviderDependentValidator{
-						AttributeName: "azure_private_link",
-						CloudProvider: "azure",
-					},
-				},
-			},
-			"gcp_private_service_connect": schema.SingleNestedAttribute{
-				Optional:    true,
-				Description: "The GCP Private Service Connect configuration.",
-				Attributes: map[string]schema.Attribute{
-					"enabled": schema.BoolAttribute{
-						Required:    true,
-						Description: "Whether Redpanda GCP Private Service Connect is enabled.",
-					},
-					"global_access_enabled": schema.BoolAttribute{
-						Required:    true,
-						Description: "Whether global access is enabled.",
-					},
-					"consumer_accept_list": schema.ListNestedAttribute{
-						Required:    true,
-						Description: "List of consumers that are allowed to connect to Redpanda GCP PSC (Private Service Connect) service attachment.",
-						NestedObject: schema.NestedAttributeObject{
-							Attributes: map[string]schema.Attribute{
-								"source": schema.StringAttribute{
-									Required:    true,
-									Description: "Either the GCP project number or its alphanumeric ID.",
-								},
-							},
-						},
-					},
-				},
-				Validators: []validator.Object{
-					validators.CloudProviderDependentValidator{
-						AttributeName: "gcp_private_service_connect",
-						CloudProvider: "gcp",
-					},
-				},
-			},
-			"kafka_api": schema.SingleNestedAttribute{
-				Optional:    true,
-				Description: "Cluster's Kafka API properties.",
-				Attributes: map[string]schema.Attribute{
-					"mtls": schema.SingleNestedAttribute{
-						Required:    true,
-						Description: "mTLS configuration.",
-						Attributes: map[string]schema.Attribute{
-							"enabled": schema.BoolAttribute{
-								Required:    true,
-								Description: "Whether mTLS is enabled.",
-							},
-							"ca_certificates_pem": schema.ListAttribute{
-								ElementType: types.StringType,
-								Required:    true,
-								Description: "CA certificate in PEM format.",
-							},
-							"principal_mapping_rules": schema.ListAttribute{
-								ElementType: types.StringType,
-								Required:    true,
-								Description: "Principal mapping rules for mTLS authentication. See the Redpanda documentation on configuring authentication.",
-							},
-						},
-					},
-				},
-			},
-			"http_proxy": schema.SingleNestedAttribute{
-				Optional:    true,
-				Description: "HTTP Proxy properties.",
-				Attributes: map[string]schema.Attribute{
-					"mtls": schema.SingleNestedAttribute{
-						Required:    true,
-						Description: "mTLS configuration.",
-						Attributes: map[string]schema.Attribute{
-							"enabled": schema.BoolAttribute{
-								Required:    true,
-								Description: "Whether mTLS is enabled.",
-							},
-							"ca_certificates_pem": schema.ListAttribute{
-								ElementType: types.StringType,
-								Required:    true,
-								Description: "CA certificate in PEM format.",
-							},
-							"principal_mapping_rules": schema.ListAttribute{
-								ElementType: types.StringType,
-								Required:    true,
-								Description: "Principal mapping rules for mTLS authentication. See the Redpanda documentation on configuring authentication.",
-							},
-						},
-					},
-				},
-			},
-			"schema_registry": schema.SingleNestedAttribute{
-				Optional:    true,
-				Description: "Cluster's Schema Registry properties.",
-				Attributes: map[string]schema.Attribute{
-					"mtls": schema.SingleNestedAttribute{
-						Required:    true,
-						Description: "mTLS configuration.",
-						Attributes: map[string]schema.Attribute{
-							"enabled": schema.BoolAttribute{
-								Required:    true,
-								Description: "Whether mTLS is enabled.",
-							},
-							"ca_certificates_pem": schema.ListAttribute{
-								ElementType: types.StringType,
-								Required:    true,
-								Description: "CA certificate in PEM format.",
-							},
-							"principal_mapping_rules": schema.ListAttribute{
-								ElementType: types.StringType,
-								Required:    true,
-								Description: "Principal mapping rules for mTLS authentication. See the Redpanda documentation on configuring authentication.",
-							},
-						},
-					},
-				},
-			},
-			"read_replica_cluster_ids": schema.ListAttribute{
-				ElementType: types.StringType,
-				Optional:    true,
-				Description: "IDs of clusters which may create read-only topics from this cluster.",
-			},
-		},
-	}
-}
-
 // Create creates a new Cluster resource. It updates the state if the resource
 // is successfully created.
 func (c *Cluster) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
@@ -334,13 +81,13 @@ func (c *Cluster) Create(ctx context.Context, req resource.CreateRequest, resp *
 
 	clusterReq, err := generateClusterRequest(model)
 	if err != nil {
-		resp.Diagnostics.AddError("unable to parse CreateCluster request", err.Error())
+		resp.Diagnostics.AddError("unable to parse CreateCluster request", utils.DeserializeGrpcError(err))
 		return
 	}
 
 	clResp, err := c.CpCl.Cluster.CreateCluster(ctx, &controlplanev1beta2.CreateClusterRequest{Cluster: clusterReq})
 	if err != nil {
-		resp.Diagnostics.AddError("failed to create cluster", err.Error())
+		resp.Diagnostics.AddError("failed to create cluster", utils.DeserializeGrpcError(err))
 		return
 	}
 	op := clResp.Operation
@@ -377,12 +124,12 @@ func (c *Cluster) Create(ctx context.Context, req resource.CreateRequest, resp *
 		return utils.NonRetryableError(fmt.Errorf("unhandled state %v. please report this issue to the provider developers", cluster.GetState()))
 	})
 	if err != nil {
-		resp.Diagnostics.AddError(fmt.Sprintf("failed to create cluster with ID %q", clusterID), err.Error())
+		resp.Diagnostics.AddError(fmt.Sprintf("failed to create cluster with ID %q", clusterID), utils.DeserializeGrpcError(err))
 		return
 	}
 	persist, err := generateModel(model, cluster)
 	if err != nil {
-		resp.Diagnostics.AddError("failed to generate model for state during cluster.Create", err.Error())
+		resp.Diagnostics.AddError("failed to generate model for state during cluster.Create", utils.DeserializeGrpcError(err))
 		return
 	}
 
@@ -401,7 +148,7 @@ func (c *Cluster) Read(ctx context.Context, req resource.ReadRequest, resp *reso
 			resp.State.RemoveResource(ctx)
 			return
 		}
-		resp.Diagnostics.AddError(fmt.Sprintf("failed to read cluster %s", model.ID), err.Error())
+		resp.Diagnostics.AddError(fmt.Sprintf("failed to read cluster %s", model.ID), utils.DeserializeGrpcError(err))
 		return
 	}
 
@@ -414,7 +161,7 @@ func (c *Cluster) Read(ctx context.Context, req resource.ReadRequest, resp *reso
 
 	persist, err := generateModel(model, cluster)
 	if err != nil {
-		resp.Diagnostics.AddError("failed to generate model for state during cluster.Read", err.Error())
+		resp.Diagnostics.AddError("failed to generate model for state during cluster.Read", utils.DeserializeGrpcError(err))
 		return
 	}
 	resp.Diagnostics.Append(resp.State.Set(ctx, persist)...)
@@ -432,25 +179,25 @@ func (c *Cluster) Update(ctx context.Context, req resource.UpdateRequest, resp *
 	if len(updateReq.UpdateMask.Paths) != 0 {
 		op, err := c.CpCl.Cluster.UpdateCluster(ctx, updateReq)
 		if err != nil {
-			resp.Diagnostics.AddError("failed to send cluster update request", err.Error())
+			resp.Diagnostics.AddError("failed to send cluster update request", utils.DeserializeGrpcError(err))
 			return
 		}
 
 		if err := utils.AreWeDoneYet(ctx, op.GetOperation(), 90*time.Minute, c.CpCl.Operation); err != nil {
-			resp.Diagnostics.AddError("failed while waiting to update cluster", err.Error())
+			resp.Diagnostics.AddError("failed while waiting to update cluster", utils.DeserializeGrpcError(err))
 			return
 		}
 	}
 
 	cluster, err := c.CpCl.ClusterForID(ctx, plan.ID.ValueString())
 	if err != nil {
-		resp.Diagnostics.AddError(fmt.Sprintf("failed to read cluster %s", plan.ID), err.Error())
+		resp.Diagnostics.AddError(fmt.Sprintf("failed to read cluster %s", plan.ID), utils.DeserializeGrpcError(err))
 		return
 	}
 
 	persist, err := generateModel(plan, cluster)
 	if err != nil {
-		resp.Diagnostics.AddError("failed to generate model for state during cluster.Update", err.Error())
+		resp.Diagnostics.AddError("failed to generate model for state during cluster.Update", utils.DeserializeGrpcError(err))
 		return
 	}
 	resp.Diagnostics.Append(resp.State.Set(ctx, persist)...)
@@ -472,11 +219,11 @@ func (c *Cluster) Delete(ctx context.Context, req resource.DeleteRequest, resp *
 		if utils.IsNotFound(err) {
 			return
 		}
-		resp.Diagnostics.AddError(fmt.Sprintf("failed to read cluster %s", model.ID), err.Error())
+		resp.Diagnostics.AddError(fmt.Sprintf("failed to read cluster %s", model.ID), utils.DeserializeGrpcError(err))
 		return
 	}
 
-	// call Delete on the cluser, if it's not already in progress. calling Delete on a cluster in
+	// call Delete on the cluster, if it's not already in progress. calling Delete on a cluster in
 	// STATE_DELETING_AGENT seems to destroy it immediately and we don't want to do that if we haven't
 	// cleaned up yet
 	if !(cluster.GetState() == controlplanev1beta2.Cluster_STATE_DELETING || cluster.GetState() == controlplanev1beta2.Cluster_STATE_DELETING_AGENT) {
@@ -484,7 +231,7 @@ func (c *Cluster) Delete(ctx context.Context, req resource.DeleteRequest, resp *
 			Id: clusterID,
 		})
 		if err != nil {
-			resp.Diagnostics.AddError("failed to delete cluster", err.Error())
+			resp.Diagnostics.AddError("failed to delete cluster", utils.DeserializeGrpcError(err))
 			return
 		}
 	}
@@ -512,7 +259,7 @@ func (c *Cluster) Delete(ctx context.Context, req resource.DeleteRequest, resp *
 		if utils.IsNotFound(err) {
 			return
 		}
-		resp.Diagnostics.AddError(fmt.Sprintf("failed to delete cluster %s", model.ID), err.Error())
+		resp.Diagnostics.AddError(fmt.Sprintf("failed to delete cluster %s", model.ID), utils.DeserializeGrpcError(err))
 		return
 	}
 }
